+++ Census/src/java/com/censo/controlador/vehiculo/CargarDatosVehiculosRunt.java
@@ -43,7 +43,7 @@ public class CargarDatosVehiculosRunt extends HttpServlet {
                 return;
             }
 
-            long idvehiculo = Long.parseLong(request.getParameter("id"));
+            int idvehiculo = Integer.parseInt(request.getParameter("id"));
 
             VehiculoDao vehiculoDao = new VehiculoDao();
             conex = vehiculoDao.conectar();
